Add files via upload
Hello world! 
this is my first ever project and i've worked on it for several hours. because i'm just it beginner it might have a lot of bugs and problems so let e know if you encounter any. also let me know on any ideas that might help improve the  project. 
this is a simple calculator and i'm looking to expand it in the future so i don't deem it a finished project yet.